made compatable with windows

+++ gui/server.py
@@ -1,26 +1,39 @@
 import os
+import platform
 import socket
 import threading
+
 
 class Server:
 
     def __init__(self, host, port, q):
+        
+        if platform.system() == "Windows":
+            
+            # Create a TCP/IP socket
+            self.sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+            
+            # Bind the socket to the port
+            server_address = ("127.0.0.1", 10000)
+            print('starting up on {} port {}'.format(server_address[0], server_address[1]))
+            self.sock.bind(server_address)
+            
+        else:
+            server_address = "./uds_socket"
+    
+            # Make sure the socket does not already exist
+            try:
+                os.unlink(server_address)
+            except OSError:
+                if os.path.exists(server_address):
+                    raise
+    
+            # Create a UDS socket
+            self.sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
 
-        server_address = './uds_socket'
-
-        # Make sure the socket does not already exist
-        try:
-            os.unlink(server_address)
-        except OSError:
-            if os.path.exists(server_address):
-                raise
-
-        # Create a UDS socket
-        self.sock = socket.socket(socket.AF_UNIX, socket.SOCK_STREAM)
-
-        # Bind the socket to the address
-        print('starting up on {}'.format(server_address))
-        self.sock.bind(server_address)
+            # Bind the socket to the address
+            print('starting up on {}'.format(server_address))
+            self.sock.bind(server_address)
 
         # Listen for incoming connections
         self.sock.listen(5)
